Upgraded the objects exported in the shiny RDS
Now shiny app has access not only to the diagnosands corresponding to the parameters specified in shiny_arguments, but also to the code that defines the parameters used in the body of the designer. This means users can copy and paste the code directly into their console, and the parameters will be defined exactly as they chose from the dropdown menus.